Seventh commit
Fixes all around. Got rid of exception-based rules for GTK3 because they just lead to nothing but trouble later down the road. GTK2 menubar color was wrong in a couple of programs for the darker variant. Shell has new OSD (anything that pops up in the middle of the screen) styling and has some changes I forgot about.